+++ src/git_portfolio/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Git portfolio domain models."""
