+++ src/git_portfolio/domain/gh_connection_settings.py
@@ -1,0 +1,7 @@
+"""Github connection settings model."""
+import collections
+
+
+GhConnectionSettings = collections.namedtuple(
+    "ConnectGithub", ["github_access_token", "github_hostname"]
+)
