+++ src/git_portfolio/domain/issue.py
@@ -1,0 +1,5 @@
+"""Issue model."""
+import collections
+
+
+Issue = collections.namedtuple("Issue", ["title", "body", "labels"])
