+++ src/git_portfolio/domain/pull_request.py
@@ -1,0 +1,17 @@
+"""Pull request model."""
+import collections
+
+PullRequest = collections.namedtuple(
+    "PullRequest",
+    [
+        "title",
+        "body",
+        "labels",
+        "confirmation",
+        "link",
+        "inherit_labels",
+        "head",
+        "base",
+        "draft",
+    ],
+)
