+++ src/git_portfolio/domain/pull_request_merge.py
@@ -1,0 +1,7 @@
+"""Pull request merge model."""
+import collections
+
+
+PullRequestMerge = collections.namedtuple(
+    "PullRequestMerge", ["base", "head", "prefix", "delete_branch"]
+)
