+++ src/git_portfolio/use_cases/__init__.py
@@ -1,0 +1,1 @@
+"""Git portfolio use cases."""
