+++ tests/test_main.py
@@ -4,6 +4,7 @@ from unittest.mock import patch
 
 import pytest
 from click.testing import CliRunner
+from pytest_mock import MockerFixture
 
 import git_portfolio.__main__
 import git_portfolio.response_objects as res
@@ -15,10 +16,23 @@ def runner() -> CliRunner:
     return CliRunner()
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-def test_git_command_success(mock_configmanager: Mock, runner: CliRunner) -> None:
+@pytest.fixture
+def mock_config_manager(mocker: MockerFixture) -> MockerFixture:
+    """Fixture for packing CONFIG_MANAGER."""
+    return mocker.patch("git_portfolio.__main__.CONFIG_MANAGER")
+
+
+@pytest.fixture
+def mock_github_manager(mocker: MockerFixture) -> MockerFixture:
+    """Fixture for packing GithubManager."""
+    return mocker.patch("git_portfolio.github_manager.GithubManager", autospec=True)
+
+
+def test_git_command_success(
+    mock_config_manager: MockerFixture, runner: CliRunner
+) -> None:
     """It outputs success message."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
 
     @git_portfolio.__main__.main.command("test")
     @git_portfolio.__main__.git_command
@@ -30,10 +44,11 @@ def test_git_command_success(mock_configmanager: Mock, runner: CliRunner) -> Non
     assert "success message" in result.output
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-def test_git_command_execute_error(mock_configmanager: Mock, runner: CliRunner) -> None:
+def test_git_command_execute_error(
+    mock_config_manager: MockerFixture, runner: CliRunner
+) -> None:
     """It calls a command an error response."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
 
     @git_portfolio.__main__.main.command("test")
     @git_portfolio.__main__.git_command
@@ -45,8 +60,7 @@ def test_git_command_execute_error(mock_configmanager: Mock, runner: CliRunner) 
     assert "Error: some error msg" in result.output
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-def test_git_command_no_repos(mock_configmanager: Mock, runner: CliRunner) -> None:
+def test_git_command_no_repos(mock_config_manager: Mock, runner: CliRunner) -> None:
     """It outputs no repos selected error message."""
 
     @git_portfolio.__main__.main.command("test")
@@ -54,19 +68,18 @@ def test_git_command_no_repos(mock_configmanager: Mock, runner: CliRunner) -> No
     def _() -> res.ResponseFailure:
         return res.ResponseFailure.build_system_error("some error msg")
 
-    mock_configmanager.config.github_selected_repos = []
+    mock_config_manager.config.github_selected_repos = []
     result = runner.invoke(git_portfolio.__main__.main, ["test"], prog_name="gitp")
 
     assert result.output.startswith("Error: no repos selected.")
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.git_use_case.GitUseCase", autospec=True)
+@patch("git_portfolio.use_cases.git_use_case.GitUseCase", autospec=True)
 def test_checkout_success(
-    mock_git_use_case: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_git_use_case: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It calls checkout with master."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
     runner.invoke(git_portfolio.__main__.main, ["checkout", "master"], prog_name="gitp")
 
     mock_git_use_case.return_value.execute.assert_called_once_with(
@@ -74,13 +87,12 @@ def test_checkout_success(
     )
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.git_use_case.GitUseCase", autospec=True)
+@patch("git_portfolio.use_cases.git_use_case.GitUseCase", autospec=True)
 def test_pull_success(
-    mock_git_use_case: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_git_use_case: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It calls pull."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
     runner.invoke(git_portfolio.__main__.main, ["pull"], prog_name="gitp")
 
     mock_git_use_case.return_value.execute.assert_called_once_with(
@@ -88,13 +100,12 @@ def test_pull_success(
     )
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.git_use_case.GitUseCase", autospec=True)
+@patch("git_portfolio.use_cases.git_use_case.GitUseCase", autospec=True)
 def test_push_success(
-    mock_git_use_case: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_git_use_case: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It calls push with origin master."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
     runner.invoke(
         git_portfolio.__main__.main, ["push", "origin", "master"], prog_name="gitp"
     )
@@ -104,13 +115,12 @@ def test_push_success(
     )
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.git_use_case.GitUseCase", autospec=True)
+@patch("git_portfolio.use_cases.git_use_case.GitUseCase", autospec=True)
 def test_status_success(
-    mock_git_use_case: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_git_use_case: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It calls status."""
-    mock_configmanager.config.github_selected_repos = ["staticdev/omg"]
+    mock_config_manager.config.github_selected_repos = ["staticdev/omg"]
     runner.invoke(git_portfolio.__main__.main, ["status"], prog_name="gitp")
 
     mock_git_use_case.return_value.execute.assert_called_once_with(
@@ -118,10 +128,8 @@ def test_status_success(
     )
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
 def test_config_init(
-    mock_github_manager: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_github_manager: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It creates pm.GithubManager."""
     runner.invoke(git_portfolio.__main__.configure, ["init"], prog_name="gitp")
@@ -129,13 +137,11 @@ def test_config_init(
     mock_github_manager.assert_called_once()
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER", autospec=True)
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
 def test_config_repos_success(
-    mock_github_manager: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_github_manager: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It call config_repos from pm.GithubManager."""
-    mock_configmanager.config_is_empty.return_value = False
+    mock_config_manager.config_is_empty.return_value = False
     result = runner.invoke(
         git_portfolio.__main__.configure, ["repos"], prog_name="gitp"
     )
@@ -144,13 +150,11 @@ def test_config_repos_success(
     assert result.output == "gitp successfully configured.\n"
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER", autospec=True)
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
 def test_config_repos_do_not_change(
-    mock_github_manager: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_github_manager: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It does not change config file."""
-    mock_configmanager.config_is_empty.return_value = False
+    mock_config_manager.config_is_empty.return_value = False
     mock_github_manager.return_value.config_repos.return_value = None
     result = runner.invoke(
         git_portfolio.__main__.configure, ["repos"], prog_name="gitp"
@@ -160,10 +164,8 @@ def test_config_repos_do_not_change(
     assert "gitp successfully configured.\n" not in result.output
 
 
-@patch("git_portfolio.__main__.CONFIG_MANAGER")
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
 def test_config_repos_no_config(
-    mock_github_manager: Mock, mock_configmanager: Mock, runner: CliRunner
+    mock_github_manager: Mock, mock_config_manager: Mock, runner: CliRunner
 ) -> None:
     """It returns error message."""
     mock_github_manager.return_value.config_repos.return_value = None
@@ -173,33 +175,69 @@ def test_config_repos_no_config(
     assert "Error" in result.output
 
 
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
-def test_create_issues(mock_github_manager: Mock, runner: CliRunner) -> None:
-    """It call create_issues from pm.GithubManager."""
+@patch(
+    "git_portfolio.use_cases.gh_create_issue_use_case.GhCreateIssueUseCase",
+    autospec=True,
+)
+def test_create_issues(
+    mock_gh_create_issue_use_case: Mock,
+    mock_github_manager: MockerFixture,
+    mock_config_manager: MockerFixture,
+    runner: CliRunner,
+) -> None:
+    """It executes gh_create_issue_use_case."""
     runner.invoke(git_portfolio.__main__.create, ["issues"], prog_name="gitp")
-    mock_github_manager.assert_called_once()
-    mock_github_manager.return_value.create_issues.assert_called_once()
+    manager = mock_github_manager.return_value
+
+    mock_gh_create_issue_use_case(manager).execute.assert_called_once()
 
 
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
-def test_create_prs(mock_github_manager: Mock, runner: CliRunner) -> None:
-    """It call create_pull_requests from pm.GithubManager."""
+@patch(
+    "git_portfolio.use_cases.gh_create_pr_use_case.GhCreatePrUseCase",
+    autospec=True,
+)
+def test_create_prs(
+    mock_gh_create_pr_use_case: Mock,
+    mock_github_manager: MockerFixture,
+    mock_config_manager: MockerFixture,
+    runner: CliRunner,
+) -> None:
+    """It executes gh_create_pr_use_case."""
     runner.invoke(git_portfolio.__main__.create, ["prs"], prog_name="gitp")
-    mock_github_manager.assert_called_once()
-    mock_github_manager.return_value.create_pull_requests.assert_called_once()
+    manager = mock_github_manager.return_value
+
+    mock_gh_create_pr_use_case(manager).execute.assert_called_once()
 
 
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
-def test_merge_prs(mock_github_manager: Mock, runner: CliRunner) -> None:
-    """It call merge_pull_requests from pm.GithubManager."""
+@patch(
+    "git_portfolio.use_cases.gh_merge_pr_use_case.GhMergePrUseCase",
+    autospec=True,
+)
+def test_merge_prs(
+    mock_gh_merge_pr_use_case: Mock,
+    mock_github_manager: MockerFixture,
+    mock_config_manager: MockerFixture,
+    runner: CliRunner,
+) -> None:
+    """It executes gh_merge_pr_use_case."""
     runner.invoke(git_portfolio.__main__.merge, ["prs"], prog_name="gitp")
-    mock_github_manager.assert_called_once()
-    mock_github_manager.return_value.merge_pull_requests.assert_called_once()
+    manager = mock_github_manager.return_value
+
+    mock_gh_merge_pr_use_case(manager).execute.assert_called_once()
 
 
-@patch("git_portfolio.github_manager.GithubManager", autospec=True)
-def test_delete_branches(mock_github_manager: Mock, runner: CliRunner) -> None:
+@patch(
+    "git_portfolio.use_cases.gh_delete_branch_use_case.GhDeleteBranchUseCase",
+    autospec=True,
+)
+def test_delete_branches(
+    mock_gh_delete_branch_use_case: Mock,
+    mock_github_manager: MockerFixture,
+    mock_config_manager: MockerFixture,
+    runner: CliRunner,
+) -> None:
     """It call delete_branches from pm.GithubManager."""
     runner.invoke(git_portfolio.__main__.delete, ["branches"], prog_name="gitp")
-    mock_github_manager.assert_called_once()
-    mock_github_manager.return_value.delete_branches.assert_called_once()
+    manager = mock_github_manager.return_value
+
+    mock_gh_delete_branch_use_case(manager).execute.assert_called_once()
