+++ tests/use_cases/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for git portfolio use cases."""
